+++ src/app/tasks/task/task.component.ts
@@ -1,7 +1,8 @@
-import {Component, EventEmitter, Input, Output} from '@angular/core';
+import {Component, EventEmitter, inject, Input, Output} from '@angular/core';
 import {Task} from "./task.model";
 import {DatePipe} from "@angular/common";
 import {CardComponent} from "../../shared/card/card.component";
+import {TasksService} from "../tasks.service";
 
 @Component({
   selector: 'app-task',
@@ -18,11 +19,14 @@ export class TaskComponent {
 
   @Output() complete = new EventEmitter<string>();
 
+  private tasksService = inject(TasksService);
+
   /**
    * na przycisk uruchamiamy onCompleteTask, który rzuca event do komponentu tasks
    */
   onCompleteTask() {
-    this.complete.emit(this.task.id);
+    // this.complete.emit(this.task.id); - zamiast emitowania eventu używamy serwisu
+    this.tasksService.removeUserTask(this.task.id);
   }
 
 }